enhancement: 도서 비즈니스 로직 테스트코드의 검증 방법을 결과값 확인을 통해 검증하도록 변경
